add or remove hive node freely using a screwdriver

+++ monkestation/code/modules/blueshift/benos/hivenode_hijacker.dm
@@ -1,10 +1,9 @@
 /obj/item/assembly/flash/xenomorph
-	name = "Hivenode Hijacker"
-	desc = "A flash device, that, when a xenomorph hive node is added, allows the user to hijack xenomorph hiveminds."
+	name = "Hivenode Hijacker (Empty)"
+	desc = "Experimental flash device with a slot for a xenomorph hive node."
 	icon_state = "flash" //need CUSTOM ICON
-	light_color = LIGHT_COLOR_PINK
+	light_color = COLOR_PURPLE
 	var/contains_node = FALSE //does NOT contain node by default
-	burnt_out = TRUE //burnt out by default, unusable without a node
 
 /obj/item/assembly/flash/xenomorph/attack(mob/living/M, mob/user)
 	if(!try_use_flash(user))
@@ -23,13 +22,35 @@
 		M.mind.add_antag_datum(xeno_datum)
 
 
-/obj/item/assembly/flash/xenomorph/attackby(obj/item/organ/internal/alien/hivenode/H, mob/living/user)
-	if(contains_node)
-		to_chat(user, "The hijacker already contains a hive node")
-	else
-		contains_node = TRUE
-		burnt_out = FALSE
-		qdel(H) //GET IT OUT OF HERE
-		to_chat(user, "You sucessfully insert the hive node into the hijacker")
+/obj/item/assembly/flash/xenomorph/attackby(obj/item/H, mob/living/user)
+	if(istype(H, /obj/item/organ/internal/alien/hivenode)) //if its a hive node
+		if(contains_node)
+			to_chat(user, "The hijacker already contains a hive node")
+		else
+			update_icon(ALL, TRUE)
+			contains_node = TRUE
+			qdel(H) //GET IT OUT OF HERE
+			to_chat(user, "You sucessfully insert the hive node into the hijacker")
+			name = "Hivenode Hijacker (Loaded)"
+			desc = "An experimental flash device used to hijack hiveminds and brainwash Xenmorphs."
+	/*
+	else if(istype(H, /obj/item/screwdriver)) //if its a screwdriver
+		if(contains_node)
+			update_icon(ALL, TRUE)
+			contains_node = FALSE //remove the node
+			to_chat(user, "You sucessfully remove the hive node")
+			name = "Hivenode Hijacker (Empty)"
+			desc = "Experimental flash device with a slot for a xenomorph hive node."
+			new /obj/item/organ/internal/alien/hivenode(user.loc)
+	*/
 	return
 
+/obj/item/assembly/flash/xenomorph/screwdriver_act(mob/living/user, obj/item/I)
+	if(contains_node)
+		update_icon(ALL, TRUE)
+		contains_node = FALSE //remove the node
+		to_chat(user, "You sucessfully remove the hive node")
+		name = "Hivenode Hijacker (Empty)"
+		desc = "Experimental flash device with a slot for a xenomorph hive node."
+		new /obj/item/organ/internal/alien/hivenode(user.loc)
+	return
